[refactor] Invert file structure so new code is top-level
https://bitwarden.atlassian.net/jira/software/projects/SG/boards/34?selectedIssue=SG-355

We are rewriting and restructuring a lot of code to make it easier to develop as a team.
As a way to seperate new and old, new code started being put in `app/modules`, with old code still existing in `app`.
There is a sentiment for that to be inverted so new code is in `app` and old code is grouped together in a subfolder.

To acheive this, the following changes were made:

1. Find a shared home for old components
	1.1. Supply the components that already exist in `app/components` with folders for their components. This created `app/components/nested-checkbox`, `./organization-switcher`, `./password-reprompt`, `./password-strength`, and `./premium-badge`.
	1.2. Move all top level component folders in `app` into `app/components`
	1.3. Move `app/modules/loose-components.module` to `app/components` and rename it to just `components.module`
2. Move all folders and files in `app/modules` to `app`
3. Cleanup. Imports needed to be adjusted across components, both in web and in bitwarden_license