[CL-62] Fix Content Tab Keyboard Navigation (#3944)
* [CL-62] Add missing modules to Dialog Service story

The IconButtonModule and SharedModule need to be available for the service to properly open the dialog.

Also fix type error for dialogSize attribute

* [CL-62] Add new tabbed dialog service story

- Update StoryDialogComponent to support different content components and button text for re-use in multiple stories
- Update the story module metadata to include Tabs and FormsField modules for the new tab story
- Add StoryTabbedDialogComponent that has tabbed content with input fields which provide tabbing targets
- Add storybook actions to provide an example of getting a result from the dialog service

* [CL-62] Remove tab panel tabIndex from tab group component

The tabIndex attribute broke keyboard navigation in Firefox and is only required on the tab labels.

* [CL-62] Introduce contentTabIndex input for bit-tab

contentTabIndex provides an interface for setting the tabPanel's tabIndex so that the tabPanel is still included in the tab sequence of the page in case it has no focusable content of its own

* [CL-62] Add tab keyboard navigation story

* Revert "[CL-62] Add new tabbed dialog service story"

This reverts commit e19216f0310fe0fdbb827be25db7aafc15d95f2f.