[EC-828] Access selector layout bugs (#4301)
* [EC-828] fix: permission column offset

* [EC-828] fix: focus border width

* [EC-828] feat: add border on hover

Border matches the hover border for the icon button.

* [EC-828] fix: properly align permission column

Chrome adds extra padding to select elements and the only way to remove it is using `appearence: none`. Unfortunately Firefox does not do this, meaning that we have different behavior when trying to use some of the built in select styles.

* [EC-828] feat: re-add select chevron

chevron is removed when setting `appearence: none`. We now have the different chevrons on a single screen thought...

* [EC-828] fix: chevron looking off-center in chrome

* [EC-828] fix: multi-select height

Min-height seems like a very hacky solution but I think we need to properly go through these styles when we have more time. Would be nice if we could change the chevron to be the same everywhere for example.

* [EC-828] fix: multi-select csp issues