[EC-887] Fix Managers can see options to edit/delete Collections they aren't assigned to (#4395)
* [EC-887] Introduce 'assigned' property to collection admin view/response

The 'assigned' property is set by the server to indicate that the collection has been explicitly assigned to the acting user. Can be used to determine if the collection can be modified/deleted by managers.

* [EC-887] Update logic to show/hide collection vault controls

Only show checkbox and ellipsis button for collections the user has access to delete and/or edit. Otherwise, hide them to avoid confusion or allowing the user to attempt actions they do not have permission to.

* [EC-887] Add missing permission message visibility property

* [EC-887] Add missing permission message to template

* [EC-887] Check for null id for the 'unassigned' collection